+++ semwork/vernead2/src/Plugins/Plugin_Rotate.cpp
@@ -17,16 +17,16 @@ void Plugin_Rotate::processImage( Image & img ) const {
     double cosf = cos(radians);
 
     // rotation
-    for (int x = 0; x < (int)img.getWidth(); x++) {
-        for (int y = 0; y < (int)img.getHeight(); y++) {
+    for (size_t x = 0; x < img.getWidth(); x++) {
+        for (size_t y = 0; y < img.getHeight(); y++) {
 
-            double xf = x - (int)mXCenter;
-            double yf = y - (int)mYCenter;
+            double xf = (double)x - mXCenter;
+            double yf = (double)y - mYCenter;
 
-            int xx = (+xf * cosf - yf * sinf + (int)mXCenter);
-            int yy = (+xf * sinf + yf * cosf + (int)mYCenter);
+            size_t xx = (+xf * cosf - yf * sinf + mXCenter);
+            size_t yy = (+xf * sinf + yf * cosf + mYCenter);
 
-            if (xx >= 0 && xx < (int)imgCopy->getWidth() && yy >= 0 && yy < (int)imgCopy->getHeight()) {
+            if (xx >= 0 && xx < imgCopy->getWidth() && yy >= 0 && yy < imgCopy->getHeight()) {
                 img.Pixel(x,y) = imgCopy->Pixel(xx, yy);
             }
             else
