+++ semwork/vernead2/src/Plugins/Plugin_Rotate.h
@@ -7,8 +7,8 @@
 #include <string>
 #include <utility>
 
-#include "Image/Image.h"
-#include "dataTypes.h"
+#include "../Image/Image.h"
+#include "../dataTypes.h"
 #include "Plugin.h"
 
 /** Image rotation using nearest neighbor interpolation
@@ -16,11 +16,11 @@
 class Plugin_Rotate: public Plugin{
 public:
 
-    explicit Plugin_Rotate(int x, int y, double direction): Plugin("Rotation"), mXCenter( x), mYCenter( y), mDirection( direction) { }
+    explicit Plugin_Rotate( ssize_t x, ssize_t y, double direction): Plugin( "Rotation"), mXCenter( x), mYCenter( y), mDirection( direction) { }
     void processImage( Image & img ) const override;
 
 protected:
 private:
-    size_t mXCenter, mYCenter;
+    ssize_t mXCenter, mYCenter;
     double mDirection;
 };
